Application/requirements.txt
absl-py==0.13.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.3
apturl==0.5.2
argcomplete==2.0.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.3
command-not-found==0.3
cryptography==38.0.4
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
executing==2.0.1
fasteners==0.14.1
filelock==3.8.0
Flask==2.0.1
frozenlist==1.4.0
future==0.18.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
grpcio-tools==1.50.0
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==3.2
ipython==8.12.3
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.1
keyring==18.0.1
kiwisolver==1.3.1
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.7
monotonic==1.5
multidict==6.0.4
netifaces==0.10.4
numpy==1.21.1
oauthlib==3.1.1
olefile==0.46
openai==0.27.8
opencv-python==4.5.3.56
packaging==21.3
pandas==1.3.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2022.10.12
pipx==1.1.0
platformdirs==2.5.2
prompt_toolkit==3.0.47
protobuf==3.17.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycosat==0.6.3
pycparser==2.22
pycups==1.9.73
pygame==2.4.0
Pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2021.1
pyusb==1.2.1
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.7.1
scour==0.37
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
soupsieve==2.4.1
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.9.0+rocm4.2
torchvision==0.10.0+rocm4.2
tqdm==4.62.0
traitlets==5.14.3
typing-extensions==3.10.0.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
usb-creator==0.3.7
userpath==1.8.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.1
xkit==0.0.0
yarl==1.9.2
